Safari-611.1.10/LayoutTests/editing/deleting/delete-list-items-in-table-cell-4.html

<!DOCTYPE html>
<html>
<body>
    <div id="container" contenteditable="true"><table border="1"><tr><td><ul><li id="li"><br></li></ul></td></tr></table></div>
    <script src="../../resources/dump-as-markup.js"></script>
    <script>
        Markup.description('When deleting the last unordered list item in a table cell we must delete the whole ordered list:');

        getSelection().collapse(document.getElementById("li"), 0);
        Markup.dump('container', 'Before');

        document.execCommand("Delete");
        Markup.dump('container', 'After');
    </script>
</body>
</html>
