LayoutTests/editing/deleting/delete-list-items-in-table-cell-4-expected.txt

When deleting the last unordered list item in a table cell we must delete the whole ordered list:

Before:
| <table>
|   border="1"
|   <tbody>
|     <tr>
|       <td>
|         <ul>
|           <li>
|             id="li"
|             <#selection-caret>
|             <br>

After:
| <table>
|   border="1"
|   <tbody>
|     <tr>
|       <td>
|         <div>
|           <#selection-caret>
|           <br>
